+++ src/orgs/smartsaver/content/survey/feedbackSurveyContent.js
@@ -1,4 +1,4 @@
-import processMarkdown from '../../../libs/marked'
+import processMarkdown from '../../../../libs/marked'
 
 const feedbackSurveyContent = {
   intro: processMarkdown(`
