+++ src/orgs/smartsaver/content/survey/thankYouSectionContent.js
@@ -1,4 +1,4 @@
-import processMarkdown from '../../../libs/marked'
+import processMarkdown from '../../../../libs/marked'
 
 const thankYouSectionContent = {
   intro: processMarkdown(`
