+++ src/orgs/smartsaver/sections/survey/Navbar.js
@@ -1,8 +1,8 @@
 import React, { Fragment } from 'react'
-import Navbar from '../../../components/Navbar'
-import NavbarItem from '../../../components/Navbar/NavbarItem'
-import NavbarDropdown from '../../../components/Navbar/NavbarDropdown'
-import brandLogo from '../images/logo-small.png'
+import Navbar from '../../../../components/Navbar'
+import NavbarItem from '../../../../components/Navbar/NavbarItem'
+import NavbarDropdown from '../../../../components/Navbar/NavbarDropdown'
+import brandLogo from '../../images/logo-small.png'
 import './Navbar.css'
 
 const SmartsaverNavbar = () => {
